+++ app/main/routes.py
@@ -10,7 +10,6 @@ from app.models import User, Post
 from app.main import bp
 
 
-
 @bp.before_request
 def before_request():
     """
@@ -22,10 +21,9 @@ def before_request():
         db.session.commit()
 
 
-
-@bp.route('/', methods=['GET', 'POST'])
-@bp.route('/index', methods=['GET', 'POST'])
-@login_required 
+@bp.route("/", methods=["GET", "POST"])
+@bp.route("/index", methods=["GET", "POST"])
+@login_required
 def index():
     """
     Route for index page
@@ -36,53 +34,86 @@ def index():
         current_app.logger.debug(f"{post}")
         db.session.add(post)
         db.session.commit()
-        flash('Your post is now live!')
-        return redirect(url_for('main.index'))
+        flash("Your post is now live!")
+        return redirect(url_for("main.index"))
 
-    posts = current_user.posts.all()
-    return render_template("index.html", title='Home Page', form=form,
-                           posts=posts)
-
+    posts = current_user.followed_posts().all()
+    return render_template("index.html", title="Home Page", form=form, posts=posts)
 
 
-@bp.route('/explore')
+@bp.route("/explore")
 @login_required
 def explore():
     """
     Route for explore
     """
     posts = Post.query.order_by(Post.timestamp.desc()).all()
-    return render_template('index.html', title='Explore', posts=posts)
+    return render_template("index.html", title="Explore", posts=posts)
 
 
-
-@bp.route('/user/<username>')
+@bp.route("/user/<username>")
 @login_required
 def user(username):
     """
     Route for user
     """
     user_ = User.query.filter_by(username=username).first_or_404()
-    posts = current_user.posts.all()
-    return render_template('user.html', user=user_, posts=posts)
+    posts = user_.posts.all()
+    return render_template("user.html", user=user_, posts=posts)
 
 
-
-@bp.route('/edit_profile', methods=['GET', 'POST'])
+@bp.route("/edit_profile", methods=["GET", "POST"])
 @login_required
 def edit_profile():
     """
     Route for editing user profile
     """
     form = EditProfileForm(current_user.username)
-    if form.validate_on_submit(): #pylint: disable=no-else-return
+    if form.validate_on_submit():  # pylint: disable=no-else-return
         current_user.username = form.username.data
         current_user.about_me = form.about_me.data
         db.session.commit()
-        flash('Your changes have been saved.')
-        return redirect(url_for('main.edit_profile'))
-    elif request.method == 'GET':
+        flash("Your changes have been saved.")
+        return redirect(url_for("main.edit_profile"))
+    elif request.method == "GET":
         form.username.data = current_user.username
         form.about_me.data = current_user.about_me
-    return render_template('edit_profile.html', title='Edit Profile',
-                           form=form)
+    return render_template("edit_profile.html", title="Edit Profile", form=form)
+
+
+@bp.route("/follow/<username>")
+@login_required
+def follow(username):
+    """
+    Follow a User
+    """
+    user_ = User.query.filter_by(username=username).first()
+    if user_ is None:
+        flash(f"User {username} not found.")
+        return redirect(url_for("index"))
+    if user_ == current_user:
+        flash("You cannot follow yourself!")
+        return redirect(url_for("user", username=username))
+    current_user.follow(user_)
+    db.session.commit()
+    flash(f"You are following {username}!")
+    return redirect(url_for("main.user", username=username))
+
+
+@bp.route("/unfollow/<username>")
+@login_required
+def unfollow(username):
+    """
+    Unfollow a User
+    """
+    user_ = User.query.filter_by(username=username).first()
+    if user is None:
+        flash(f"User {username} not found.")
+        return redirect(url_for("index"))
+    if user_ == current_user:
+        flash("You cannot unfollow yourself!")
+        return redirect(url_for("user", username=username))
+    current_user.unfollow(user_)
+    db.session.commit()
+    flash(f"You are not following {username}.")
+    return redirect(url_for("main.user", username=username))